Dev/chn/bandwidth (#318)
* ldr-hdr,
low precision for shaders

* shadow distance per scene

* fix game-builds & disable double buffering

* fix windows build

* hdr pipeline change in editor

* conjunct fxaa gamma and tonemapping pass to one pass

* fix memory leak

* fix android build

* remove glm submodule

* track glm 0.9.9.8

* fix android build window

* set android build window volatile window

* fix gizmo calculation when object is in the middle

* fix material window & gammaToneMapFxaa pass

* remove debug code

---------

+++ Resources/Engine/Shaders/gBufferFrag.shader
@@ -17,7 +17,7 @@
 	<!--
 		#version 300 es
 		precision highp float;
-		precision highp int;
+		precision lowp int;
 
 		in vec3 v_pos;
 		in vec3 v_normal;
