+++ Resources/Engine/Shaders/gBufferFrag_alphamask.shader
@@ -6,7 +6,7 @@
 	<!--
 		#version 300 es
 		precision highp float;
-		precision highp int;
+		precision lowp int;
 
 	-->
 	</source>
